First step towards avoiding to re-download an old issue.
Fors this we create a new directory to put the issues in.
When we are about to download a new one, we first check if
it already exists. If it does, we simply open up the existing
file. If it doesn't exist, we delete all old issues to preserve
space, and the proceed as usual to download and show the
current issue.

Still left to do: Parsing the file name of the issue from the
current ePaper web service. Right now we use the timestamp, which
of course always causes the issue to be re-downloaded.